% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asmBUSCOcheck.R
\name{asmBUSCOcheck}
\alias{asmBUSCOcheck}
\title{Look for missing BUSCO in the haplotig assembly}
\usage{
asmBUSCOcheck(
  asm1.busco = NULL,
  asm2.busco = NULL,
  summary = FALSE,
  out = NULL
)
}
\arguments{
\item{summary}{summarise the number of BUSCO genes that are found in the subset of contigs from the secondary/haplotig assembly (optional; logical)}

\item{out}{supplying a output prefix will write a headerless, single-column file of contig names from the secondar/haplotig assembly. File name will be in the format [out].txt (optional, string)}

\item{asm.busco1}{dataframe containing the full_table.tsv or full_table.txt output from BUSCO or compelasm for the main/primary assembly (required)}

\item{asm.busco2}{dataframe containing the full_table.tsv or full_table.txt output from BUSCO or compelasm for the secondary/haplotig assembly (required)}
}
\description{
Look for missing BUSCO in the haplotig assembly
}
\details{
Determine if there are BUSCO genes missing in the main assembly that are
are found in the alternate assembly. BUSCO should be run on two sister
assemblies—that is they should be output from a single set of reads from the
same genome assembler as main and alternate assemblies OR output from a
haplotig/duplicate contig purging program such as purge_haplotigs or
purge_dups. Function requires two BUSCO (or compleasm) full_table.tsv files.
}
