% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/busco.scaffold.filter.R
\name{busco.scaffold.filter}
\alias{busco.scaffold.filter}
\title{Generate a filter list of scaffolds}
\usage{
busco.scaffold.filter(jbat.list = jbat.list, busco.dat = busco.dat, out = NULL)
}
\arguments{
\item{jbat.list}{a list object generated from the read.jbat.review() function (REQUIRED)}

\item{busco.dat}{a dataframe generated from the read.busco() function (REQUIRED)}

\item{out}{a string that will be used as output file prefixes (default = NULL; optional)}
}
\description{
Generate a filter list of scaffolds
}
\details{
This creates a filter list and a summary table of scaffolds which will output
tab-delimited txt files. The main goal of this function is to create two lists which will help to subset
a scaffold assembly to contain only main scaffolds and unplaceable contigs/scaffolds. The generated lists
and their contents are:
 (out_)scaffold_summary.txt
   sequence: name of the sequence, scaffold_1:n where n is the total number of scaffolds
   jbat_status: either "scaffold" or "debris". when editing HiC contact maps in Juicebox
                some sequences may be designated as "debris", which are typically
                sequenced duplicates. fragments are considered scaffolds.
   length: length of scaffold derived from taking the sum of all contigs that comprise
            that scaffold
   complete: the number of complete single (S) copy BUSCOs found on a sequence
   dup: the number of complete duplicated (D) BUSCOs found on a sequence
   frag: the number of fragmented BUSCOs (F) found on a sequence
   keep: TRUE/FALSE was the scaffold placed in the keep list? this determination
         is made by a set of criteria:
         for TRUE:
            1) the sequence contain any complete or fragmented BUSCOs (regardless of JBAT status)
            2) the sequence is considered a scaffold and does not contain any BUSCOs
            3) the sequence contains duplicate BUSCOs that cannot be found on any
               other sequence
         for FALSE:
            1) the sequence contains only duplicate BUSCOs and these BUSCOs are
               also located on sequences in TRUE
            2) the jbat_status is considered "debris" and does not contain any BUSCOs

(out_)list_of_scaff_to_keep.txt
   a single column list without headers containing sequence names of scaffolds
   that should be kept. these are sequences where keep is TRUE in (out_)scaffold.summary.txt

(out_)debris_and_dups_scaffolds.txt
   a single column list without headers containing sequence names of scaffolds
   that contain only BUSCO duplicates and debris. these are sequences where
   keep is FALSE in (out_)scaffold.summary.txt

The user can feed either list into their fasta handling program such as
seqtk subseq or samtools faidx to create a subsetted fasta containing only
the sequences included in either list.
}
