% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insp_stbed2df.R
\name{insp_stbed2df}
\alias{insp_stbed2df}
\title{Convert structural_error.BED output from Inspector into a dataframe}
\usage{
insp_stbed2df(path_to_bed, summarise = T)
}
\arguments{
\item{path_to_bed}{The absolute file path to the *_structural_error.BED file.
The \* should be a some sort of ID string}

\item{summarise}{Logical (T/F) default: summarise = T. The contents of structural_error.bed has lists each error as a row with columns
for the start and end of positions of the error, the number of reads that support that the error is there, the type of error, and the size of the error (in bp).
summarising will calculate the size (in bp) of each type of error for the assembly.}
}
\description{
The python program Inspector (available from: https://github.com/Maggi-Chen/Inspector)
can be used to identify structural errors (>50 bp) in hifi genome assemblies.
This function takes the structural_error.BED output and converts it into
a dataframe for easy manipulation in R.
}
\details{
As of this writing (Feb. 3, 2023), Inspector does not name its output files with identifying information (by design).
For use with this function, the file must be copied and renamed with identifying information separated by underscores (_).
For example: specimen_assembly_structural_error.BED
Note that if you need to use inspector-correct, then the files cannot be renamed (hence copying and renaming).

If there are multiple structural_error.bed files and the goal is to summarise each file,
it may be best if this function is used in lapply (see example). This will generate a list of dataframes
which can made into a single dataframe using bind_rows (in dplyr) or do.call and rbind (see example). To do this, first create
a vector of absolute paths to each structural_error.bed file or aggregate all structural_error.bed
files into a single directory and make that the working directory. Then, use
list.files() in lapply with this function to generate a list of dataframes with summarised
error lengths. Finally, use bind_rows on this list object to create a single dataframe
with each row representing an assembly.
}
\examples{
#write example output of a structural_error.bed output from Inspector to the
#current working directory
cat(`Nmex03_ipa-np_insp_structural_error.bed`,
file = "Nmex03_ipa-np_insp_structural_error.bed")
cat(`Nmex03_ipa-np_pd_insp_structural_error.bed`,
file = "Nmex03_ipa-np_pd_insp_structural_error.bed")

#for a single file, summarised
insp_stbed2df("Nmex03_ipa-np_insp_structural_error.bed", summarise = T)

#for a single file, unsummarised output
insp_stbed2df("Nmex03_ipa-np_insp_structural_error.bed", summarise = F)

#for multiple files, summarised. For unsummarised output, simply set summarise = F
do.call('rbind', lapply(list.files(pattern = '.bed', insp_stbed2df, summarise = T)))
}
\author{
Gen Morinaga
}
